tcp: stop all timers when in tcp_close()
instead of doing that in tcp_discardcb().  A connection in CLOSED state
shall not need any timers.  Assert that no timer is rescheduled after that
with tcp_timer_activate().  This will fix a panic observed after
f4574e2dc5a4 just by not running the tcp_timer_rexmt() on a closed
connection.  That connection still remains in an invalid state, where
socket send buffer is empty, but TCP thinks there is stuff to send.  So
the root cause of the panic isn't fixed.