tcp: short circuit tcp_output() on a TCPS_CLOSED connection
In tcp_output() if flags have RST, that means we are in TCPS_CLOSED
state, called from tcp_drop().  This tcp_output() shall only send
RST and do not call tcp_timer_activate() neither we need to update
any bits of tcpcb, as it is already on the way for garbage collection.