callout: provide CALLOUT_TRYLOCK flag
If a callout was initialized with the flag, then the callout(9) system
will not drop the callwheel lock in softclock_call_cc() to obtain the
callout lock.  Instead it will use try lock semantic to obtain the
callout's lock.  In case of a failure the callout will be rescheduled
to the nearest cycle reusing the original precision value.  The main
benefit of such behavior is not the avoidance of the lock contention
in the callout thread, but the fact that callout with such flag can
be actually stopped in a safer manner.

XXXGL: writeup

Caveat: in this e