unix/stream: fixups on top of 2024 version
This shall be squashed into the main new unix(4) commit, with proper
ammending the commit message.

Changes accumulated so far:

- provide own pr_sopoll

This version would check into the peers buffer if connected.  It uses
protocol lock to check for SOLISTENING(so), which is safe since conversion
uses both locks.

- provide own pr_kqfilter

- try to fix uio problem with aio(4) on the side of the unix(4)

This may be moved to side of aio(4).

aio: aio needs wakeup on the send buffer, hack around that

The problem with socket aio(4) is that it uses socket buffer locks for
queueing and piggybacking on this locking calls soreadable() and
sowriteable() directly.  Ideally it should use pr_sopoll() method.

At this point solve the problem on the side of unix(4), but ideally
it should be socket aio(4) problem (not a first class citizen) than
problem of unix(4), definitely a primary facility.