aio: aio needs wakeup on the send buffer, hack around that
The problem with socket aio(4) is that it uses socket buffer locks for
queueing and piggybacking on this locking calls soreadable() and
sowriteable() directly.  Ideally it should use pr_sopoll() method.

At this point solve the problem on the side of unix(4), but ideally
it should be socket aio(4) problem (not a first class citizen) than
problem of unix(4), definitely a primary facility.

This shall be squashed into the main new unix(4) commit.